Mu.Servant.Server

Description

This module allows you to serve a Mu Server as an OpenAPI Swagger REST end-point. In particular, it translates to the kind of type-level APIs used by Servant.

Convert Mu to Servant

servantServerHandlers

Arguments

:: forall pname m chn ss handlers. (ServantServiceHandlers ('Package pname ss) m chn ss handlers, ExtraFor ('Package pname ss) ~ EmptyAPI)

=> (forall a. m a -> Handler a)
how to turn the inner Mu monad into Handler, use toHandler (or a composition with it) in most cases

-> ServerT chn () ('Package pname ss) m handlers
server to be converted

-> Server (PackageAPI ('Package pname ss) handlers)

Converts a Mu server into Servant server by running all Mu handler actions in the Handler type. This version assumes no additional routes in the Servant server when compared to Mu's.

servantServerHandlersExtra

Arguments

:: forall pname m chn ss handlers. ServantServiceHandlers ('Package pname ss) m chn ss handlers

=> (forall a. m a -> Handler a)
how to turn the inner Mu monad into Handler, use toHandler (or a composition with it) in most cases

-> Server (ExtraFor ('Package pname ss))
additional handler for the extra route

-> ServerT chn () ('Package pname ss) m handlers
server to be converted

-> Server (PackageAPI ('Package pname ss) handlers)

Converts a Mu server into Servant server by running all Mu handler actions in the Handler type. This version should be used when additional routes have been added in the Servant version.

toHandler :: ServerErrorIO a -> Handler a

Reinterprets a Mu server action as a Servant handler.

packageAPI :: ServerT chn t pkg s handlers -> Proxy (PackageAPI pkg handlers)

Obtains a Servant API Proxy value for use with functions like serve and layout.

swagger :: forall pname ss handlers chn m. HasSwagger (ServicesAPI ('Package pname ss) ss handlers) => ServerT chn () ('Package pname ss) m handlers -> Swagger

Converts the information from a Mu server into a Swagger document.

Required annotations

data ServantRoute

ServantRoute represents the URL path components of a route. It is used as an AnnotatedPackage domain to override the default path for a Method. When used in an AnnService, the specified TopLevelRoute is used as a prefix for all Methods in that Service. 1. List of components for the route, 2. HTTP method which must be used, 3. HTTP status code of a successful HTTP response from a specific Method. Use 200 for the usual status code.

Constructors

ServantAdditional Type

ServantTopLevelRoute [Symbol]

ServantRoute [Symbol] StdMethod Nat

type DefaultServantContentTypes = 'ServantContentTypes '[JSON] ('Just ('ServantStreamContentType NewlineFraming JSON))

data ServantContentTypes

ServantContentTypes represents that acceptable content types that can be used when a message in encoded: 1. in a unary (non-streaming) HTTP request/response body, 2. encoded in a streaming HTTP request/response body. It is used as an AnnotatedSchema domain.

Constructors

ServantContentTypes

Fields

unary :: [Type]

stream :: Maybe ServantStreamContentType

data ServantStreamContentType

Constructors

ServantStreamContentType

Fields

framing :: Type

streamContentType :: Type

data StdMethod

Constructors

GET

POST

HEAD

PUT

DELETE

TRACE

CONNECT

OPTIONS

PATCH